Refactor to include the io routines od ppymech in the io module
