+++ tests/not_with_pytest/sem.py
@@ -8,7 +8,7 @@ import numpy as np
 # Import relevant modules
 from pynektools.interpolation.mesh_to_mesh import PRefiner
 from pynektools.interpolation.point_interpolator.single_point_legendre_interpolator import LegendreInterpolator as element_interpolator_c
-from pynektools.ppymech.neksuite import preadnek
+from pynektools.io.ppymech.neksuite import preadnek
 from pynektools.datatypes.msh import Mesh as msh_c
 from pynektools.datatypes.coef import Coef as coef_c
 from pynektools.datatypes.field import Field as field_c
