+++ tests/not_with_pytest/tensor_sem.py
@@ -19,7 +19,7 @@ import numpy as np
 # Import relevant modules
 from pynektools.interpolation.mesh_to_mesh import PRefiner
 from pynektools.interpolation.point_interpolator.single_point_legendre_interpolator import LegendreInterpolator as element_interpolator_c
-from pynektools.ppymech.neksuite import preadnek
+from pynektools.io.ppymech.neksuite import preadnek
 from pynektools.datatypes.msh import Mesh as msh_c
 
 def test_sem_and_tensor_sem(n_new = 8, elem = 1, max_pts = 1, use_torch = False, verbose = False):
